+++ TFGVolandoVoy/TFGVolandoVoy.csproj
@@ -58,14 +58,6 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <AndroidResource Remove="Resources\Images\retos.png" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <MauiImage Remove="Resources\Images\retos.png" />
-	</ItemGroup>
-
-	<ItemGroup>
 	  <None Remove="Resources\AppIcon\logo_app.png" />
 	  <None Remove="Resources\Fonts\Oswald-Bold.ttf" />
 	  <None Remove="Resources\Fonts\Oswald-ExtraLight.ttf" />
@@ -75,14 +67,10 @@
 	  <None Remove="Resources\Fonts\Oswald-SemiBold.ttf" />
 	  <None Remove="Resources\Images\invisible.png" />
 	  <None Remove="Resources\Images\invisibledark.png" />
-	  <None Remove="Resources\Images\retos.png" />
+	  <None Remove="Resources\Images\retoss.png" />
 	  <None Remove="Resources\Images\visible.png" />
 	  <None Remove="Resources\Images\visibledark.png" />
 	  <None Remove="Resources\Splash\splash_vl.png" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <Content Include="Resources\Images\retos.png" />
 	</ItemGroup>
 
 	<ItemGroup>
@@ -99,6 +87,9 @@
 	<ItemGroup>
 	  <Compile Update="Vista\DetallesLocalidadVtn.xaml.cs">
 	    <DependentUpon>DetallesLocalidadVtn.xaml</DependentUpon>
+	  </Compile>
+	  <Compile Update="Vista\RetosVnt_.xaml.cs">
+	    <DependentUpon>RetosVnt_.xaml</DependentUpon>
 	  </Compile>
 	  <Compile Update="Vista\VntPrincipal.xaml.cs">
 	    <DependentUpon>VntPrincipal.xaml</DependentUpon>
