--- TFGVolandoVoy/Vista/Retos.xaml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="TFGVolandoVoy.Vista.Retos"
-             Title="Retos">
-
-    <ScrollView>
-        <VerticalStackLayout>
-            <Label Text="Retos de la localidad"
-                   FontSize="Large"
-                   VerticalOptions="Center"
-                   HorizontalOptions="Center"/>
-            <Image Source="retos.png"
-                   HeightRequest="25"
-                   WidthRequest="100"
-                   Aspect="AspectFit"
-                   x:Name="retos_titulo"/>
-            <!--<ListView>
-                <ListView.ItemTemplate>
-                    <DataTemplate>
-                        <ViewCell>
-                            <Frame CornerRadius="10" Padding="5">
-                                <StackLayout Orientation="Vertical" Padding="20,0" Margin="0,5">
-                                    <Label Text="{Binding NombreLocalidad}" VerticalOptions="Center"/>
-                                    <Label Text="{Binding TextoReto}" VerticalOptions="Center"/>
-                                </StackLayout>
-                            </Frame>
-                        </ViewCell>
-                    </DataTemplate>
-                </ListView.ItemTemplate>
-            </ListView>-->            
-        </VerticalStackLayout>
-    </ScrollView>
-</ContentPage>
